fix: hice el componente ImportantNote mas reutilizable
