+++ src/index.scss
@@ -20,3 +20,4 @@
 /* SCSS Components Files */
 @import './views/Layout/Layout.scss';
 @import './components/index.scss';
+@import './components/ImportantNote/ImportantNote.scss';
